[fix] renderData app.js로부터 분리 및 안 쓰는 파일 들 정리
