+++ assets/data/menuLinkText.js
@@ -183,3 +183,21 @@ export const specialMenuLinkText = [
     alt: "출석체크",
   },
 ];
+
+export const appDownLoad = {
+  triggerId: "dropdown-download-trigger",
+  triggerText: "배민찬 앱 다운로드",
+  dropdownId: "dropdown-download",
+  dropdown: [
+    {
+      text: "앱스토어",
+      url: "#",
+      className: "appstore",
+    },
+    {
+      text: "구글플레이스토어",
+      url: "#",
+      className: "playstore",
+    },
+  ],
+};
