--- assets/data/tabBtnText.js
@@ -1,26 +1,0 @@
-export const tabBtnText = [
-  {
-    text: "오늘은 뭐먹지",
-    url: "#",
-  },
-  {
-    text: "풍성한 고기반찬",
-    url: "#",
-  },
-  {
-    text: "바다향가득 반찬",
-    url: "#",
-  },
-  {
-    text: "추천해요 아이반찬",
-    url: "#",
-  },
-  {
-    text: "할인특가 세트상품",
-    url: "#",
-  },
-  {
-    text: "맛있는 간식타임",
-    url: "#",
-  },
-];
