+++ style/sass/_bestSection.scss
@@ -153,4 +153,9 @@
     white-space: nowrap;
     text-align: left;
   }
+  .badge-box {
+    position: absolute;
+    left: 0;
+    bottom: 137px;
+  }
 }
